First pass version of the SDK generator which supports exposing PlayFab to Lua using behavior context.
Allows Lua to create request objects and send request over the PlayFab bus. Allows Lua to listen for notifications using the notification bus (currently this bus is not yet complete as it currently sends all response messages on channel 0 rather than using a ticket system).

To do and known issues:
- map and vector<enum> types not yet reflected.
- MultitypeVar not yet reflected.
- Boxed<T> types don't yet have automatic conversion to/from T, so are a little more difficult to use from Lua.
- Need to set-up a global notification bus (which listens to all requests) and an addressed bus which listens for responses to specific requests.
- Lua requires requests to be sent without using any optional parameters. E.g:
  PlayFabClientRequestBus.Broadcast.GetFriendsList(request,nil,nil,nil)
- Reflected enum names are not final